requirement.txt
cryptography==2.9.2
decorator==4.4.2
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pycryptodome==3.9.7
PyJWT==1.7.1